Game-field/fix: diveded game-field drawing into modules and created a new one for hero implementation. Refused to use background class, instead - just drawing bgImage directly
